Updated openstack/openstack Project: openstack-infra/system-config  91bca8d574e94b5fce6d88acb5ff7730b8790b41
Remove mlocate from static.o.o

mlocate is filling up the disk trying to index logs and docs.

Its default is not to index remote mounts, but since these are mounted
on static.o.o as block devices it descends into them.  Another option
is to update PRUNEPATHS in updatedb.conf, but since this is wholly
unnecessary let's KISS and just get rid of it.

It is only installed because it is a "suggests" of findutils, so has
no reverse-dependencies.

Change-Id: Ib23f3f1fb3397b66f897a0d284da521ce50293e8