Updated openstack/openstack Project: openstack/requirements  953a6f36370439c5a59ceb1cf1f1e99181b742bc
Add pypi package 'docker' to requirements

The pypi package 'docker-py' [1] has been renamed to 'docker' [2].
This patch add 'docker' to requirements as a first step to handle
the renaming. Later, we needs to work with all projects that
has reference to 'docker-py' and help them to migrate to the
new package.

[1] https://pypi.python.org/pypi/docker-py
[2] https://pypi.python.org/pypi/docker

Change-Id: Ib66a803653c5c0b92c981612f3788f5ab36dda0d