Updated openstack/openstack Project: openstack/neutron-lbaas  831743f0ee18601d86f7755424f56bff3a430f6b
Keep user_group option of haproxy section in lbaas_agent.ini

The default user_group is nogroup and it works well in ubuntu,
but it makes error in centos whose defualt group should be nobody
and we can't find the user_group option in the config file of
lbaas_agent.ini. We should preserve this option.

Change-Id: I220bc5dc360cefabccbc4653465319c79f964e18
Closed-Bug: #1650355