Updated openstack/openstack Project: openstack/tripleo-heat-templates  854b5120cc04418b34e45a523d0ea67ae64a448e
Describe use of Manila/CephFS in capabilites_map

This is meant to describe in the UI how to use the
manila-cephfsnative-config environment file to deploy Manila with
Ceph as backend using either a TripleO managed or an externally
managed Ceph cluster.

Also adds a puppet-ceph.yaml environment file meant to enable the
deployment of a Ceph cluster, given that storage-environment.yaml
is meant to be customized by the user instead.

Change-Id: Iefd7056a9bc079025e6ac4dd50edcd2e2635e1b0