Updated openstack/openstack Project: openstack/tripleo-heat-templates  681fdb50e06158d38f53cb281b1a3042fe3de929
Reduce memcached memory configuration

Previously the memcached configuration was set to use the defaults which
would be 95% of the avaiable ram in the system. This can lead to memory
contention issues if memcache is heavily utilized. This change reduces
the default to 50% and exposes the ability to tune this configuration.

Change-Id: Ie8a48ff4cf509e93d7c1487813d5feed5e5131a4
Closes-Bug: #1662941