Updated openstack/openstack Project: openstack/horizon  392ddec13b3294ea08f0d02be664c55b5fd35e0b
Disentangle domain context from effective domain

Since the existence of a domain token was equivalent with having
selected a domain context with Keystone V2, some code confuses the
two. This is no longer true for Kestone V3, so we have to separate
the two concepts and use domain context when we mean the domain
context.

Close-bug: #1661537

Change-Id: Ifa66d8c397e34d16a4534e7216eb11c752699505