Updated openstack/openstack Project: openstack/requirements  79adc2fa801fb1a1f0fd1ce692cbaa84b3a9b1db
Add kingbird client to requirements

Kingbird has released the first version of its client.
Since out intention is to use the client for tempest test
in order to get rid of ugly curl requires.
Since we would like to comply with the global openstack
requirements, we need kingbirdclient to be part of the list
so that we could add it to the requirements.txt

Change-Id: I5247ab6f0f4d9950905fb846c2c6a3ec868ec23b