Updated openstack/openstack Project: openstack/cinder  ef6c81709e8fb4ac23462b4f991eb76923d878e8
Fix SolidFire cgsnap clone bug

When I initially implemented consistency group support in the SolidFire driver
I did not reailze that a use case/path would involve creating a volume from a
snapshot that was part of a group snap/cgsnap.

Due to the way that SolidFire takes group snapshots, and the naming scheme
that is used for the resulting snapshots, some additional lookups are needed
in order to locate the correct snapshot. This was implemented correct for the
various cgsnap calls but it is missing from the create volume from clone path.
This patch corrects that and the fix was verified against a cluster running
SolidFire release Oxygen 8.0.1.3.

Closes-Bug: #1663059
Change-Id: I8d9bb674e03b55c549ab693a57711e6455eb8f93