Updated openstack/openstack Project: openstack/cinder  683258ca8292f7b366ac3f8c9ceb89d41ffc5904
Adds attached migration to multibackend-matrix

This test was left behind as it wasn't passing the LVM -> NFS
migration setup. As it wasn't possible to reproduce and fix localy,
let's add this to the gate so we can reproduce the error scenario.
Additionaly to that, some bug fixes need this to be fully tested.

Change-Id: Icd9138c1bc1335e9465184749bf6dd5733c605a3