Updated openstack/openstack Project: openstack/cinder  9c95ba8e38bb4e7a6f011bcd42b0240f577cb8da
Fix attachments after attached migration

When migrating an attached volume, after the migration completion,
the volume gots stuck and can not be deleted or detached. This
is happening due an error in the volume completion implementation.
Now, before detaching the source volume, we first save the
attachments and then re-connect then in the destination volume.

Closes-bug: #1659914
Change-Id: Ie9a7815df99869b62209d99bfdcf51b250a8dba3