Updated openstack/openstack Project: openstack-infra/project-config  5bdb005be2ac2208a786918a3c40ba81df750e6b
add tripleo non-voting job into zaqar gate

With the motivation of adding more diversity in how do we test
zaqar, this patch will add a tripleo job, non-voting.

The idea is to bring more feedback when testing patches in these
projects and detect what failures we could have outside devstack.

The jobs are not voting. We don't expect zaqar team to open TripleO job
logs. The idea is to let TripleO folks know when a job is red so they
can investigate and tell if something critical is broken in the patch.

Change-Id: Iee53a6d9aba7252c0edf8a4a816839c766d35391