Updated openstack/openstack Project: openstack-infra/project-config  df4f52a01f736358870784d1c3f2ce80ef4be51f
add tripleo non-voting job into congress gate

With the motivation of adding more diversity in how do we test
congress, this patch will add a tripleo job, non-voting.

The idea is to bring more feedback when testing patches in these
projects and detect what failures we could have outside devstack.

The jobs are not voting. We don't expect Congress team to open TripleO job
logs. The idea is to let TripleO folks know when a job is red so they
can investigate and tell if something critical is broken in the patch.

Change-Id: Ie1f72470fddc3626ef915effb2535847b0a02217