Updated openstack/openstack Project: openstack-infra/project-config  031caed06909837a0075464b2bbe2d8de35475c3
tripleo-ci move multinode-upgrades-nv to check queue

Having this in the experimental queue has allowed some regressions to
creep in, as the job is highly coupled to code in the tripleo-ci repo.

So move it to the check queue (still non voting for now) such that we
hopefully spot breaking changes before they merge.

Change-Id: I1d5a970dd653b97b8b494bf450543d07fbfd83d3