Updated openstack/openstack Project: openstack-infra/project-config  b48ea40d6b04d6712e74378ac3113132007587fa
neutron: Run in-tree Neutron tempest tests with concurrency 2

These tests use Ubuntu image for instances which makes it longer to
boot. On certain gate resources it takes sometimes longer than 6
minutes to succesfully start sshd on instances.

This patch sets the tempest concurrency run to 2 to avoid
overloading the environment.

Change-Id: I197ac2764cf9b87e0544b5655fb5db8a1dcba555