Updated openstack/openstack Project: openstack-infra/project-config  eb8a00ef00651e91c9c7cd6351e9138377048301
Add openstack-ansible non-voting jobs to ARA check queue

The ARA implementation in the openstack-ansible gate was recently
broken.
In https://review.openstack.org/#/c/435486/ we're making it so we
can leverage basic OSA func jobs as a way of testing that each
ARA patch doesn't break OSA.

The jobs are non-voting for now, we might consider making them voting
if they are reliable.

Change-Id: I65812dabf9f62d688a85dc01275e5aa16235a09a
Depends-On: Ia7b46424740da94b82c57b828b53c7993fdaa76b