Updated openstack/openstack Project: openstack/nova  c746974054e1a78032051fef7202a1b98dc9603f
Raise correct error instead of class exist in Placement API

When create resource class using placement API, and at
the DB layer, we have a retry logic for 100 loop, and
error will raise when the retry limit exceed, currently
a log about limit exceed is generated but a
ResourceClassExists Error is raised, which is a little
bit weird and might confuse the users.

We'd better add a particular Exception for this kind of
scenario and it could be used in other places also.

Co-Authored-By: huangtianhua@huawei.com
Closes-Bug: #1658868

Change-Id: I89156be82ad520bdeacb12b31b44e7170b77ce3f