Updated openstack/openstack Project: openstack-dev/grenade  7f7def83cf7fbc35ae1a97e983e5fc0cba6e8d88
Run the nova-status upgrade check before starting upgraded services

This simply runs the nova-status upgrade check command, which was
introduced in Ocata, after we've run upgrade scripts but before
starting services with new code. This should let us know that we
successfully setup the deployment for running the new code after
the upgrade.

Depends-On: I1e4dae17cf9d1336bf0ca72948b135b02434ba15

Change-Id: I554e0df221ca6c37ad15d0fa87b7a335f674a7af