Updated openstack/openstack Project: openstack/cinder  2decc903c4bbf9cde8b2483c77b63bdea0084f22
create snapshot with generic group fail with XtremIO Driver.

fix a bug in the xtremio driver which cause a create
snapshot from generic group with volumes to fail. the bug was that
the driver did not expect to have group_id instead of
consistency_group_id as part of the cgsnapshots object.

Closes-bug: #1664209
Change-Id: I5682fa5eaa7f265c7e10a2af5704847129704637