Updated openstack/openstack Project: openstack/neutron  9979acf85924093feb8aaf1a1e1c647eaba5a5d1
Make ML2 OVO push notification asynchronous

The OVO push notification logic was blocking the AFTER_UPDATE
event for all core resources. This isn't problematic for individual
HTTP API calls. However, the agent current updates the status of
every port it wires two times, including on agent restarts.

In order to drastically reduce the impact of this notifier, this
patch moves it into a spawned eventlet coroutine so it doesn't
block the main AFTER_UPDATE events. A semaphore is used to prevent
multiple coroutines from trying to hit the database at once in the
background.

This doesn't change the semantics of the notifier since its goal
is always to send the latest events.

Partial-Bug: #1665215
Change-Id: Ic259bad6f65f4bc45f5b900e0979c9a91865089b