Updated openstack/openstack Project: openstack/neutron  e96580c8fd468d26b0baea2e5096d7664c3b9834
Avoid segment DB lookup in _expand_segment on port

The port context already has a network attached to it that contains
all of the segments for that network. In all cases the segments the
port wants to lookup should be present here so we can just iterate
over them without a separate DB query.

Partial-Bug: #1665215
Change-Id: I9014c4e401134f4db908ed989e0095d0ca95b7c9