Updated openstack/openstack Project: openstack/tripleo-heat-templates  b0928df1515a50ab26d7b62fc666bd6b3701c028
Add environment for deployed-server with pacemaker

A new environment file to be used when using the deployed-server roles
data at deployed-server/deployed-server-roles-data.yaml. This ensures
the Pre and Post Puppet Tasks for the ControllerDeployedServer role are
mapped to the stacks that handle maintenance mode and resource restarts
for pacemaker on stack-update.

Change-Id: I1ca52dfb3a3b669e128ebb0a28d9e36a1807faad
Closes-Bug: #1665060