Updated openstack/openstack Project: openstack/zaqar  4bb39586683aff69bafae0d838f50829ef5431c2
Ignore conflict when deleting queues with Swift

When we delete a queue with the Swift backend, we must first delete all
the objects from the container, and then the container itself.
Unfortunately, it's not transactional, so messages could have been put
in the mean time. As queues are lazy objects, I think it's fair to
ignore the conflict error in that case.

Change-Id: I8bcbb22265c401efa9a43c5db558beca4a0134a4