Updated openstack/openstack Project: openstack/python-designateclient  8d806158ae816b9b5a561c8ef2ca83538c355a32
Simplify OSC doc structure

This page should be linked from the main docs as a place to see the
various commands that are possible. This moves the more verbose
examples to one place and delete the other page. There was also some
grammar that bugged me so I changed it

Change-Id: I8bc23e493794a43d7f6eb4d5bd5dce965c8ff11d